upload txt file with original size with ktx file

+++ scripts/image-to-ktx.sh
@@ -11,15 +11,20 @@ nextPowerOf2() {
   node getNextPower $1 $2
 }
 #nextPowerOf2 23 $WIDTH
+echo "STARTING..."
+mkdir "${STAGED_FOLDER}/ktx"
 
 let end=0
 getImageCount () {
+  mkdir "${STAGED_FOLDER}/ktx/sizes"
   for FILE in "$1"/*
     do
       if [ -d $FILE ]
       then
         getImageCount $FILE
       else
+        NAME=`file $FILE | cut -d ',' -f 1 | cut -d '.' -f 1 | cut -d '/' -f 5`
+        touch "./${STAGED_FOLDER}/ktx/sizes/${NAME}.txt"
         let end=$end+1
       fi
   done
@@ -28,7 +33,6 @@ getImageCount "${STAGED_FOLDER}/images"
 
 let start=0
 imageToKTX () {
-  mkdir "${STAGED_FOLDER}/ktx"
   for FILE in "$1"/*
     do
       if [ -d $FILE ]
@@ -44,10 +48,13 @@ imageToKTX () {
         WIDTH=`file $FILE | cut -d ',' -f 8 | cut -d 'x' -f 1`
         HEIGHT=`file $FILE | cut -d ',' -f 8 | cut -d 'x' -f 2`
 
+        echo $WIDTH >> "${STAGED_FOLDER}/ktx/sizes/${NAME}.txt"
+        echo $HEIGHT >> "${STAGED_FOLDER}/ktx/sizes/${NAME}.txt"
+
         KTXWIDTH=`node ./scripts/nearestPower.js $WIDTH`
         KTXHEIGHT=`node ./scripts/nearestPower.js $HEIGHT`
 
-        TOKTX=`toktx --2d --genmipmap --bcmp --resize 2048x2048 --t2 --verbose "public/staging/ktx/${NAME}" $FILE`
+        TOKTX=`toktx --2d --genmipmap --bcmp --resize "${KTXWIDTH}x${KTXHEIGHT}" --t2 --verbose "public/staging/ktx/${NAME}" $FILE`
 
         let start=$start+1
         echo "${start}/${end} Images Compressed."
@@ -55,25 +62,28 @@ imageToKTX () {
   done
   echo " "
   let start=0
+  touch $RESULT_FILE
   for FILE in "$STAGED_FOLDER/ktx"/*
     do
-      touch $RESULT_FILE
+      if [ -f $FILE ]
+      then
+        NAME=`file $FILE | cut -d ',' -f 1 | cut -d '.' -f 1 | cut -d '/' -f 4`
+        DATE_S=`date +%s`
+        CLOUD_FOLDER="${NAME}-${DATE_S}"
+        CLOUD_URL="https://d27rt3a60hh1lx.cloudfront.net/textures/${CLOUD_FOLDER}/${NAME}.ktx2"
 
-      NAME=`file $FILE | cut -d ',' -f 1 | cut -d '.' -f 1 | cut -d '/' -f 4`
-      DATE_S=`date +%s`
-      CLOUD_FOLDER="${NAME}-${DATE_S}"
-      CLOUD_URL="https://d27rt3a60hh1lx.cloudfront.net/textures/${CLOUD_FOLDER}/${NAME}.ktx2"
+        aws s3 cp "${FILE}" "s3://spaces-gallery-assets/textures/${CLOUD_FOLDER}/${NAME}.ktx2" || handle_error
+        aws s3 cp "${STAGED_FOLDER}/ktx/sizes/${NAME}.txt" "s3://spaces-gallery-assets/textures/${CLOUD_FOLDER}/${NAME}.txt" || handle_error
 
-      aws s3 cp "${FILE}" "s3://spaces-gallery-assets/textures/${CLOUD_FOLDER}/${NAME}.ktx2" || handle_error
+        echo $NAME >> $RESULT_FILE
+        echo $CLOUD_URL >> $RESULT_FILE
+        echo " " >> $RESULT_FILE
 
-      echo $NAME >> $RESULT_FILE
-      echo $CLOUD_URL >> $RESULT_FILE
-      echo " " >> $RESULT_FILE
-
-      echo "Image successfully uploaded! Compressed Texture can be found at -> ${CLOUD_URL}"
-      let start=$start+1
-      echo "${start}/${end} Textures Uploaded. 🤛👁👄👁🤜"
-      echo " "
+        echo "Image successfully uploaded! Compressed Texture can be found at -> ${CLOUD_URL}"
+        let start=$start+1
+        echo "${start}/${end} Textures Uploaded. 🤛👁👄👁🤜"
+        echo " "
+      fi
   done
   rm -rf "${STAGED_FOLDER}/ktx"
 }
